Divide RandomForestCounterfactualMilp class in sub-classes
